fix: simplify dependency resolver edge cases
