+++ crates/pcb-zen/src/lsp/mod.rs
@@ -457,7 +457,7 @@ impl LspEvalContext {
         }
 
         let mut resolution = crate::get_workspace_info(&self.file_provider, &workspace_root)
-            .and_then(|ws| crate::resolve_workspace_dependencies(ws, file_path, false, false))
+            .and_then(|ws| crate::resolve_workspace_dependencies(ws, &workspace_root, false, false))
             .unwrap_or_else(|_| ResolutionResult::empty());
         resolution.canonicalize_keys(&*self.file_provider);
         let resolution = Arc::new(resolution);
